+++ dynamodb/docs.go
@@ -1,0 +1,4 @@
+/*
+Package dynamodb contains an implementation of the `gokv.Store` interface for Amazon DynamoDB.
+*/
+package dynamodb